Add menu item for every L&F, fix L&F pref saving
Upon GUI creation (in the Menu factory), create one menu item for each of the
available L&Fs on the system.
When changing the L&F, save the L&F class name into the preference manager
instead of the `LookAndFeelStrategy` descendant class's name; accordingly, when
initializing the GUI, use that class name and pass it to the flexible L&F
setter class.