Move item-command mapping into addMenuItem()
Create a secondary version of `addMenuItem()` (overloading the existing one)
which includes adding the item's command to the `EventCommandLookUp` map, which
is the default use-case.
This reduces code-duplication a bit and makes it more readable at the same
time.
The old, overloaded version of the method persists, since there are some
non-default usages, especially in regards to radio and check boxes.