Fix a minor bug for the case for a wrong L&F class
The commit fixes a bug that occurs, if an invalid class name is stored
in the local preferences. In that case an exception is thrown indicating
the L&F class could not be correctly loaded. The fix registeres the current
L&F class name in the preferences, such that the exception will not
be thrown again.